@@ -1,69 +1,55 @@
 {
-  "name": "mlcl_elements",
-  "version": "2.0.0",
-  "description": "Elements ORM ",
+  "name": "elements",
   "main": "dist/index.js",
   "scripts": {
-    "test": "mocha -R spec"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/molecuel/mlcl_elements.git"
+    "test": "mocha",
+    "cover": "istanbul cover ./node_modules/mocha/bin/_mocha && npm run remap && npm run remaphtml && npm run remaplcov",
+    "remap": "./node_modules/.bin/remap-istanbul -i coverage/coverage.json -o coverage/coverage-remap.json",
+    "remaphtml": "./node_modules/.bin/remap-istanbul -i coverage/coverage.json -o coverage/html -t html",
+    "remaplcov": "./node_modules/.bin/remap-istanbul -i coverage/coverage.json -o coverage/lcov-remap.info -t lcovonly",
+    "coveralls": "npm run cover && cat ./coverage/lcov-remap.info | coveralls"
   },
   "keywords": [],
-  "author": "Dominic Böttger <Dominic.Boettger@inspirationlabs.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/molecuel/mlcl_elements/issues"
-  },
-  "typings": "dist/index.d.ts",
-  "homepage": "https://github.com/molecuel/mlcl_elements#readme",
   "devDependencies": {
-    "@types/async": "^2.0.36",
-    "@types/del": "^2.2.31",
-    "@types/elasticsearch": "^5.0.9",
-    "@types/glob": "^5.0.30",
     "@types/gulp": "^3.8.32",
-    "@types/gulp-plumber": "0.0.29",
+    "@types/gulp-plumber": "^0.0.29",
     "@types/gulp-tslint": "^3.6.30",
-    "@types/gulp-typescript": "0.0.32",
-    "@types/lodash": "^4.14.41",
+    "@types/gulp-typescript": "^0.0.32",
+    "@types/lodash": "^4.14.42",
     "@types/merge2": "^0.3.29",
-    "@types/minimatch": "^2.0.29",
     "@types/mocha": "^2.2.33",
-    "@types/mongodb": "^2.1.36",
-    "@types/node": "0.0.0",
-    "@types/orchestrator": "0.0.30",
-    "@types/q": "0.0.32",
     "@types/should": "^8.1.30",
-    "@types/vinyl": "^1.2.30",
     "@types/when": "^2.4.28",
     "assert": ">=0.0.1",
-    "bson": "^0.5.2",
+    "coveralls": "^2.11.15",
     "del": "^2.2.0",
     "gulp": "^3.9.1",
+    "gulp-gh-pages": "^0.5.4",
     "gulp-plumber": "^1.1.0",
-    "gulp-sourcemaps": "^1.6.0",
+    "gulp-sourcemaps": "^2.2.2",
     "gulp-tsconfig-files": "^0.2.5",
     "gulp-tslint": "^4.3.2",
-    "gulp-typescript": "^3.0.0",
+    "gulp-typedoc": "^2.0.1",
+    "gulp-typescript": "^3.1.3",
     "gulp-util": "^3.0.7",
     "gulpclass": "0.0.8",
-    "lodash": "^4.13.1",
+    "istanbul": "^0.4.5",
     "merge2": "^1.0.1",
     "mocha": "*",
-    "should": "^11.1.2",
+    "mocha-lcov-reporter": "^1.2.0",
+    "should": ">= 0.0.1",
     "tslint": "^3.5.0",
     "tslint-stylish": "^2.1.0-beta",
-    "typescript": "^2.0.0"
+    "typedoc": "^0.5.1",
+    "typescript": "^2.1.0"
   },
   "dependencies": {
-    "@reactivex/rxjs": "^5.0.0-rc.4",
-    "elasticsearch": "^12.0.1",
-    "mongodb": "^2.1.18",
+    "@molecuel/di": "^0.0.2",
+    "@reactivex/rxjs": "^5.0.0-rc.5",
+    "config": "^1.24.0",
     "reflect-metadata": "^0.1.3",
-    "tsvalidate": ">=0.0.1",
-    "validator": "^5.4.0"
+    "remap-istanbul": "^0.8.4",
+    "tsvalidate": "0.0.10"
   },
   "devconfig": {
     "paths": {
