@@ -1,21 +1,22 @@
 'use strict';
 import 'reflect-metadata';
-import { Elements } from '../index';
-import { IElement } from '../interfaces/IElement';
-import { IValidatorError, MongoID } from 'tsvalidate';
+import { MlclElements } from './MlclElements';
+import * as TSV from 'tsvalidate';
+// import * as Interfaces from '../Interfaces';
+import { injectable } from '@molecuel/di';
 
-export class Element implements IElement {
-  @MongoID()
-  public _id: any;
-  public elements: Elements;
-  public static elements: Elements;
-  public getElements(): Elements {
-    return Element.elements;
+@injectable
+export class Element {
+  constructor(private elements: MlclElements) {}
+  @TSV.ValidateType()
+  public id: any;
+  public getFactory(): MlclElements {
+    return this.elements;
   }
-  public setFactory(elements) {
+  public setFactory(elements: MlclElements) {
     this.elements = elements;
   }
-  public validate(): IValidatorError[] {
+  public validate(): TSV.IValidatorError[] {
     return this.elements.validate(this);
   }
   public save(upsert?: boolean): Promise<any> {
