@@ -10,6 +10,8 @@ import * as del from 'del';
 import * as fs from 'fs';
 import * as merge from 'merge2';
 import * as sourcemaps from 'gulp-sourcemaps';
+import * as typedoc from 'gulp-typedoc';
+import * as ghPages from 'gulp-gh-pages';
 
 @Gulpclass()
 export class Gulpfile {
@@ -45,19 +47,34 @@ export class Gulpfile {
   }
 
   /**
+   * Clean docs directory
+   */
+  @Task('clean::docs')
+  cleandocs(cb: Function) {
+    return del(['./docs/**'], cb);
+  }
+
+  /**
+   * Clean .publish directory
+   */
+  @Task('clean::.publish')
+  cleanpubl(cb: Function) {
+    return del(['./.publish/**'], cb);
+  }
+
+  /**
    * Typescript lint task
    */
   @Task('ts::lint')
   tslint() {
-    let lintoptions: any = {
-      emitError: false,
-      sort: true,
-      bell: true
-    }
     return gulp.src(this.config.paths.source)
       .pipe(plumber())
-      .pipe(tslint())
-      .pipe(tslint.report(require('tslint-stylish'), lintoptions));
+      .pipe(tslint({
+            formatter: 'prose'
+      }))
+      .pipe(tslint.report({
+            summarizeFailureOutput: true
+        }));
   }
 
   /**
@@ -65,7 +82,7 @@ export class Gulpfile {
    */
   @Task('ts::compile')
   tscompile(): any {
-    let sourcepaths = ['typings/index.d.ts', 'typings/main.d.ts', 'typings_override/index.d.ts'];
+    let sourcepaths = [];
     sourcepaths.push(this.config.paths.source);
     let tsResult = gulp.src(sourcepaths)
       .pipe(sourcemaps.init())
@@ -74,7 +91,7 @@ export class Gulpfile {
     return merge([
       tsResult.dts.pipe(gulp.dest(this.config.paths.dist)),
       tsResult.js
-        .pipe(sourcemaps.write('.'))
+        .pipe(sourcemaps.write('.', {includeContent: false, sourceRoot: '../src'}))
         .pipe(gulp.dest(this.config.paths.dist))
     ]);
   }
@@ -93,6 +110,35 @@ export class Gulpfile {
     return merge([
       tsResult.js.pipe(gulp.dest('test/'))
     ]);
+  }
+
+  @Task('ghpages::deploy')
+  ghpagesdeploy() {
+    return gulp.src('./docs/**/*')
+      .pipe(ghPages());
+  }
+
+  @Task('docs')
+  docs() {
+    return gulp
+            .src(['./src/*.ts'])
+            .pipe(typedoc({
+            // TypeScript options (see typescript docs)
+            target: 'es6',
+            // includeDeclarations: true,
+            // Output options (see typedoc docs)
+            out: './docs',
+            mode: 'file',
+            disableOutputCheck: false,
+            gaID: 'n',
+            // TypeDoc options (see typedoc docs)
+            ignoreCompilerErrors: true
+        }));
+  }
+
+  @SequenceTask('deploy') // this special annotation using "run-sequence" module to run returned tasks in sequence
+  deploy() {
+    return ['docs', 'ghpages::deploy', 'clean::docs', 'clean::.publish'];
   }
 
   @SequenceTask('build') // this special annotation using "run-sequence" module to run returned tasks in sequence
